Added tokenStatus "flag" to prevent two weaknesses
As brought to my attention by SparkFun customer #437 in a comment on the tutorial previous code was susceptible to a reverse roll jam attack, where the attacker impersonates the base, and prevents the base from invalidating the token by sending repeated invalid signatures.

Also, the fact that it was exposed in the source code that an invalid token was all 0x00s, meant that the attacker to ask the remote to sign a null token (all 0x00s), and then use this any time they wish. Thanks #437!!!

I also moved the "destroy" token to a function called "refreshToken"

Also added some more debug messages to keep track of status.